+++ packages/markup/src/markdown/marks/index.ts
@@ -3,4 +3,5 @@
 export { codeMarkCodec } from "./code.js";
 export { emMarkCodec } from "./em.js";
 export { linkMarkCodec } from "./link.js";
+export { strikeMarkCodec } from "./strike.js";
 export { strongMarkCodec } from "./strong.js";